+++ BreweryMaster/BreweryMaster.API/Recipe/Models/RecipeResponse.cs
@@ -4,10 +4,10 @@
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public float? BLGScale { get; set; }
+        public decimal? BLGScale { get; set; }
         public int? IBUScale { get; set; }
-        public float? ABVScale { get; set; }
-        public float? SRMScale { get; set; }
+        public decimal? ABVScale { get; set; }
+        public decimal? SRMScale { get; set; }
         public string? TypeName { get; set; }
         public string? StyleName { get; set; }
     }
